Allow using native document screenshot facility
WebdriverCSS has custom code that takes a screenshot of the whole page
by stitching together multiple viewport screenshots. This is necessary
for drivers that don't natively support screenshots of the whole page.
However, this approach is slow for large pages and not necessary when
the driver in use allows native screenshots of the whole page.